requirements.txt
Authomatic==0.0.10
Brlapi==0.5.6
Django==1.7.6
GnuPGInterface==0.3.2
Mako==0.5.0
MarkupSafe==0.15
MySQL-python==1.2.3
PAM==0.4.2
PIL==1.1.7
PyICU==1.3
PyJWT==0.4.1
Pyste==0.9.10
Twisted-Core==11.1.0
Twisted-Names==11.1.0
Twisted-Web==11.1.0
adium-theme-ubuntu==0.3.2
apt-clone==0.2.1
apt-xapian-index==0.44
apturl==0.5.1ubuntu3
argparse==1.2.1
chardet==2.0.1
command-not-found==0.2.44
configglue==1.0
debtagshw==0.1
defer==1.0.2
dell-recovery==0.0.0
dirspec==3.0.0
## FIXME: could not find svn URL in dependency_links for this package:
distribute==0.6.24dev-r0
django-oauth2-provider==0.2.6.1
django-tastypie==0.12.1
duplicity==0.6.18
fluendo-eula==0.6
httplib2==0.7.2
jockey==0.9.7
keyring==0.9.2
language-selector==0.1
launchpadlib==1.9.12
lazr.restfulclient==0.12.0
lazr.uri==1.0.3
louis==2.3.0
numpy==1.6.1
nvidia-common==0.0.0
oauth==1.0.1
-e git+https://github.com/idan/oauthlib.git@fd239fca84644896b1971cf24bc6213d065adb86#egg=oauthlib-dev
onboard==0.97.1
oneconf==0.2.8.1
pexpect==2.3
piston-mini-client==0.7.2
progressbar==2.2
protobuf==2.4.1
pyOpenSSL==0.12
pycrypto==2.4.1
pycups==1.9.61
pycurl==7.19.0
pyinotify==0.9.2
pyserial==2.5
pysmbc==1.0.13
python-apt==0.8.3ubuntu7
python-dateutil==1.5
python-debian==0.1.21ubuntu1
python-distutils-extra==2.34
python-mimeparse==0.1.4
python-openid==2.2.5
python-social-auth==0.2.2
python-virtkey==0.60.0
pyxdg==0.19
reportlab==2.5
requests==2.6.0
requests-oauthlib==0.4.2
rhythmbox-ubuntuone==3.0.0
scikit-image==0.10.1
scipy==0.9.0
sessioninstaller==0.0.0
shortuuid==0.4.2
simplejson==2.3.2
six==1.9.0
software-center-aptd-plugins==0.0.0
system-service==0.1.6
ubuntuone-couch==0.3.0
ubuntuone-installer==3.0.2
ubuntuone-storage-protocol==3.0.2
ufw==0.31.1-1
unattended-upgrades==0.1
unity-lens-video==0.3.5
unity-scope-video-remote==0.3.5
usb-creator==0.2.23
wadllib==1.3.0
wsgiref==0.1.2
wxPython==2.8.12.1
wxPython-common==2.8.12.1
xdiagnose==2.5.2ubuntu0.1
xkit==0.0.0
zope.interface==3.6.1
